Refactor GitHub lists integration to use modular GithubService
